backend/requirements.txt
# FastAPI Framework
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-multipart>=0.0.5

# Azure AI Services
azure-ai-documentintelligence>=1.0.0b1
azure-search-documents>=11.4.0
azure-ai-translation-text>=1.0.0b1
azure-core>=1.29.0

# LLM Providers
google-generativeai>=0.3.0
openai>=1.0.0

# Data Processing
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# PDF Processing
pypdf2>=3.0.0

# Vector & Math
numpy>=1.24.0

# Utilities
httpx>=0.25.0
aiofiles>=23.0.0

# Security (for production)
python-jose[cryptography]>=3.3.0
